Reuse logic in path tracing scene generator
